--- __mocks__/api.electron.js
@@ -1,4 +1,0 @@
-module.exports = {
-  getDefaultLanguage: jest.fn(() => 'en'),
-  setNotFirstTimeFlag: jest.fn()
-}
