+++ app/components/about/index.js
@@ -1,9 +1,9 @@
 import { compose } from 'redux'
-import { withNamespaces } from 'react-i18next'
+import { withTranslation } from 'react-i18next'
 import { withRouter } from 'react-router-dom'
 import About from './about'
 
 export default compose(
-  withNamespaces(),
+  withTranslation(),
   withRouter
 )(About)
