+++ app/components/noteEditor/index.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 import { compose } from 'redux'
-import { withNamespaces } from 'react-i18next'
+import { withTranslation } from 'react-i18next'
 import NoteEditor from './noteEditor'
 import {
   updateNoteTitle,
@@ -20,5 +20,5 @@ const mapDispatchToProps = dispatch => ({
 
 export default compose(
   connect(mapStateToProps, mapDispatchToProps),
-  withNamespaces()
+  withTranslation()
 )(NoteEditor)
