+++ app/components/popoverIcon/index.js
@@ -1,8 +1,8 @@
 import { compose } from 'redux'
-import { withNamespaces } from 'react-i18next'
+import { withTranslation } from 'react-i18next'
 
 import PopoverIcon from './popoverIcon'
 
 export default compose(
-  withNamespaces()
+  withTranslation()
 )(PopoverIcon)
