+++ app/i18n/index.js
@@ -1,5 +1,5 @@
 import i18n from 'i18next'
-import { reactI18nextModule } from 'react-i18next'
+import { initReactI18next } from 'react-i18next'
 import { getDefaultLanguage } from '../utils/api.electron'
 import { SUPPORTED_LANGS } from './locales'
 import isDev from 'electron-is-dev'
@@ -10,7 +10,7 @@ SUPPORTED_LANGS.forEach(lang => {
   resources[lang.code] = require('./locales/' + lang.path).default
 })
 
-i18n.use(reactI18nextModule).init({
+i18n.use(initReactI18next).init({
   lng: getDefaultLanguage(),
   debug: isDev,
   resources: resources
