+++ app/reducers/note.js
@@ -1,7 +1,9 @@
 import uuidv4 from 'uuid/v4'
 import { ACTIONS } from '../constants/actions'
 import { NOTE_STATUS } from '../constants/noteStatus'
-import logger from 'electron-log'
+import { getLogger } from '../utils/api.electron'
+
+const logger = getLogger()
 
 const INITIAL_STATE = {
   activeNoteIndex: ACTIONS.NOT_SELECTED_NOTE,
