+++ app/utils/__mocks__/api.electron.js
@@ -1,0 +1,7 @@
+module.exports = {
+  getDefaultLanguage: jest.fn(() => 'en'),
+  setNotFirstTimeFlag: jest.fn(),
+  getLogger: jest.fn(() => ({
+    warn: jest.fn()
+  }))
+}
