+++ docs/List-of-dependencies.md
@@ -5,10 +5,10 @@ Packages which are used during development or which are used to build Deer's bun
 |           Name          |  Version  |    Usage   |
 | ----------------------- | --------- | ---------- |
 | @babel/core             | ^7.2.2   | Needed for other babel's packages |
-| @babel/plugin-proposal-object-rest-spread | ^7.3.1 | A Babel plugin responsible for transforming rest and spread properties for objects |
+| @babel/plugin-proposal-object-rest-spread | ^7.3.2 | A Babel plugin responsible for transforming rest and spread properties for objects |
 | @babel/preset-env       | ^7.3.1    | A Babel plugin responsible for compiling Javascript ES6 code down to ES5 |
 | @babel/preset-react     | ^7.0.0    | A Babel plugin responsible for compiling JSX down to Javascript |
-| @octokit/rest           | ^16.14.1  | GitHub REST API client for JavaScript |
+| @octokit/rest           | ^16.15.0  | GitHub REST API client for JavaScript |
 | babel-core              | 7.0.0-bridge.0 | Needed for other babel's packages (jest) |
 | babel-loader            | ^8.0.5    | A Webpack loader responsible for taking in the ES6 code and making it understandable |
 | css-loader              | ^2.0.1    | A Webpack loader responsible for collecting CSS from all the css files and put it into a string |
@@ -18,21 +18,19 @@ Packages which are used during development or which are used to build Deer's bun
 | electron-devtools-installer | ^2.2.4 | Install DevTool extensions into Electron |
 | electron-reload         | ^1.4.0    | Reload contents of when the source files are changed |
 | enzyme                  | ^3.8.0    | JavaScript Testing utilities for React |
-| enzyme-adapter-react-16 | ^1.8.0    | JavaScript Testing utilities for React |
+| enzyme-adapter-react-16 | ^1.9.1    | JavaScript Testing utilities for React |
 | file-loader             | ^3.0.1    | A Webpack loader responsible for emitting files that will be bundled and returning their public URLs (e.g. Images) |
 | glob-fs                 | ^0.1.7    | file globbing for node.js. |
-| gulp                    | ^4.0.0    | A toolkit for automating tasks  |
-| gulp-util               | ^3.0.8    | Utility functions for gulp plugins |
 | husky                   | ^1.3.1    | Prevent bad commit, push by doing checks before it takes place |
-| jest                    | ^24.0.0   | Delightful JavaScript Testing |
+| jest                    | ^24.1.0   | Delightful JavaScript Testing |
 | jquery                  | ^3.3.1    | Needed for electron-builder packages |
 | jsdoc-to-markdown       | ^4.0.1    | Generates markdown API documentation |
 | mime                    | ^2.4.0    | A comprehensive library for mime-type mapping |
 | redux-mock-store        | ^1.5.3    | A mock store for testing your redux async action creators and middleware. |
 | standard                | ^12.0.1   | Linting utility for JavaScript and JSX |
 | style-loader            | ^0.23.1   | A Webpack loader responsibile for taking the output string generated by by css-loader and put it inside the <style> tags |
-| webpack                 | ^4.29.0   | A module bundler for JavaScript files |
-| webpack-cli             | ^3.2.1    | Required by webpack |
+| webpack                 | ^4.29.3   | A module bundler for JavaScript files |
+| webpack-cli             | ^3.2.3    | Required by webpack |
 | webpack-dev-server      | ^3.1.14   | Provides live reloading for development |
 | webpack-merge           | ^4.2.1    | Provides a merge function that concatenates arrays and merges objects creating a new object |
 
@@ -44,21 +42,21 @@ Those are dependencies that are essential for software to work.
 
 |          Name         |    Version    |    Usage   |
 | --------------------- | ------------- | ---------- |
-| @material-ui/core     | ^3.9.1        | A front-end framework |
+| @material-ui/core     | ^3.9.2        | A front-end framework |
 | @material-ui/icons    | ^3.0.2        |  Material Design Svg Icons converted to Material-UI React components |
 | electron-is-dev       | ^1.0.1        | Check if Electron is running in development |
-| electron-log          | ^3.0.0        | A multi-transport async logging library |
+| electron-log          | ^3.0.1        | A multi-transport async logging library |
 | electron-store        | ^2.0.0        | Save and load user preferences |
 | electron-window-state | ^5.0.3        | Store and restore window sizes and positions |
-| i18next               | ^13.1.2       | internationalization framework |
+| i18next               | ^15.0.1       | internationalization framework |
 | jss                   | ^9.8.7        | A lib for generating Style Sheets with JavaScript. |
 | jss-rtl               | ^0.2.3        | Enables right-to-left support by flipping every rule |
 | pouchdb-browser       | ^7.0.0        | A pocket-sized database. |
-| prop-types            | ^15.6.2       | A Runtime type checking for React props |
-| react                 | ^16.7.0       | A JavaScript library for building user interfaces |
+| prop-types            | ^15.7.1       | A Runtime type checking for React props |
+| react                 | ^16.8.1       | A JavaScript library for building user interfaces |
 | react-custom-scrollbars | ^4.2.1      | React scrollbars component |
-| react-dom             | ^16.7.0       | React package for working with the DOM |
-| react-i18next         | ^9.0.10       | Internationalization for react |
+| react-dom             | ^16.8.1       | React package for working with the DOM |
+| react-i18next         | ^10.0.4       | Internationalization for react |
 | react-jss             | ^8.6.1        | JSS integration with React |
 | react-quill           | ^1.3.3        | The Quill rich-text editor as a React component |
 | react-redux           | ^6.0.0        | React bindings for Redux |
