+++ jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  setupTestFrameworkScriptFile: '<rootDir>/setupTest.js',
+  setupFilesAfterEnv: ['<rootDir>/setupTest.js'],
   testPathIgnorePatterns: ['<rootDir>/setupTest.js'],
   moduleNameMapper: {
     '\\.(css|less|sass|scss)$': '<rootDir>/__mocks__/styleMock.js',
