+++ package.json
@@ -12,18 +12,15 @@
   "bugs": {
     "url": "https://github.com/abahmed/Deer/issues"
   },
-  "electronVersion": "4.0.1",
+  "electronVersion": "4.0.4",
   "license": "MIT",
-  "copyright": "© 2018, Deer",
+  "copyright": "© 2019, Deer",
   "private": true,
   "scripts": {
     "build-prod": "webpack --mode=production",
     "deploy": "node deploy.js",
-    "dev": "webpack-dev-server --mode=development",
-    "docs": "gulp docs",
-    "electron": "electron .",
-    "lint": "standard app/**/*.{js,jsx} --fix",
-    "postinstall": "electron-builder install-app-deps",
+    "docs": "node genDocs.js",
+    "lint": "standard app/**/*.{js,jsx} *.js --fix",
     "release": "npm run build-prod && electron-builder --publish never",
     "release:linux": "npm run build-prod && electron-builder --linux ",
     "release:linux:deb": "npm run build-prod && electron-builder --linux deb",
@@ -31,60 +28,57 @@
     "release:linux:zip": "npm run build-prod && electron-builder --linux zip",
     "release:mac": "npm run build-prod && electron-builder --mac --publish never",
     "release:win": "npm run build-prod && electron-builder --win --publish never",
-    "start": "gulp start",
+    "start": "webpack-dev-server --mode=development",
     "test": "jest --verbose"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2",
-    "@babel/plugin-proposal-object-rest-spread": "^7.3.1",
+    "@babel/plugin-proposal-object-rest-spread": "^7.3.2",
     "@babel/preset-env": "^7.3.1",
     "@babel/preset-react": "^7.0.0",
-    "@octokit/rest": "^16.14.0",
+    "@octokit/rest": "^16.15.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.5",
     "css-loader": "^2.1.0",
     "devtron": "^1.4.0",
-    "electron": "^4.0.3",
+    "electron": "^4.0.4",
     "electron-builder": "^20.38.5",
     "electron-devtools-installer": "^2.2.4",
     "electron-reload": "^1.4.0",
     "enzyme": "^3.8.0",
-    "enzyme-adapter-react-16": "^1.8.0",
+    "enzyme-adapter-react-16": "^1.9.1",
     "file-loader": "^3.0.1",
     "glob-fs": "^0.1.7",
-    "gulp": "^4.0.0",
-    "gulp-util": "^3.0.8",
     "husky": "^1.3.1",
-    "jest": "^24.0.0",
+    "jest": "^24.1.0",
     "jquery": "^3.3.1",
     "jsdoc-to-markdown": "^4.0.1",
     "mime": "^2.4.0",
     "redux-mock-store": "^1.5.3",
     "standard": "^12.0.1",
     "style-loader": "^0.23.1",
-    "webpack": "^4.29.0",
-    "webpack-cli": "^3.2.1",
+    "webpack": "^4.29.3",
+    "webpack-cli": "^3.2.3",
     "webpack-dev-server": "^3.1.14",
     "webpack-merge": "^4.2.1"
   },
   "dependencies": {
-    "@material-ui/core": "^3.9.1",
+    "@material-ui/core": "^3.9.2",
     "@material-ui/icons": "^3.0.2",
     "electron-is-dev": "^1.0.1",
-    "electron-log": "^3.0.0",
+    "electron-log": "^3.0.1",
     "electron-store": "^2.0.0",
     "electron-window-state": "^5.0.3",
-    "i18next": "^14.0.1",
-    "immutable": "^4.0.0-rc.12",
+    "i18next": "^15.0.1",
     "jss": "^9.8.7",
     "jss-rtl": "^0.2.3",
     "pouchdb-browser": "^7.0.0",
-    "prop-types": "^15.6.2",
-    "react": "^16.7.0",
+    "prop-types": "^15.7.1",
+    "react": "^16.8.1",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^16.7.0",
-    "react-i18next": "^9.0.10",
+    "react-dom": "^16.8.1",
+    "react-i18next": "^10.0.4",
     "react-jss": "^8.6.1",
     "react-quill": "^1.3.3",
     "react-redux": "^6.0.0",
@@ -92,7 +86,6 @@
     "redux": "^4.0.1",
     "redux-actions": "^2.6.4",
     "redux-thunk": "^2.3.0",
-    "sqlite3": "^4.0.6",
     "uuid": "^3.3.2"
   },
   "husky": {
